requirements.txt
PyYAML>=6.0
pytest>=7.0.0
pytest-mock>=3.8.0
